+++ CHANGELOG.md
@@ -5,6 +5,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 See [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and [Angular Commit Guidelines](https://github.com/angular/angular.js/blob/master/DEVELOPERS.md#commits)
 for guidelines followed by this project.
 
+## [Unreleased]
+
+### Updated
+
+- Validate only Legal Entities
+- Validate only 3rd and 4th generation
+
 ## [0.1.0] - 2023-03-23
 
 ### Added